wikipedia_pageviews_airflow/requirements.txt
requests
apache-airflow-providers-postgres
psycopg2-binary